Don't use live data for EditText errors because they should not persist across states
